Add a hack to fix ticket #288.
We do this by ensuring that the type_, which can be one of ['security',
'bugfix', 'enhancement', 'newpackage'], is a standard utf8-encoded byte string.
When this comes in as a unicode string, it somehow causes problem.  I'm not
sure how or why, but this hack "fixes" this issue.  Adding unicode strings to
the PackageUpdate's type EnumCol does not fix it.