Ticket #286: Show an update example
When displaying the update, parse the MarkDown syntax in the update
notes to render nicely formatted HTML.

This is a follow-up on commit e9318aa03cd2771fd4c4af42bd552ca465e6f3bd that
introduced a new dependency on the markdown2 module. The spec file needed to be
updated for the new Requires, which was forgotten in the previous commit.