Don't set the package.suggest_requires permanently when the developer enables it.
By doing this, we have accumulated a large list unnecessary of packages that
now suggest that the user reboot their machine.  If a package has ever been a
part of a large multi-build update where the developer has enabled the 'Suggest
Reboot' option, this package would always have this option.  This is not ideal.
By disabling this, the 'suggest reboot' metadata option will only be enabled if
the update has a package on bodhi's hardcoded list, or if the developer enables
that option manually.