Have bodhi start managing the 'pending' koji tags.
When an update is initially requested to be pushed to testing, bodhi
will add a dist-fNN-updates-testing-pending tag to the builds.  From
here, AutoQA will automatically handle mashing repositories of these
pending builds to perform various tests on.

Once requested to go to the stable repository, bodhi will then add the
dist-fNN-updates-pending tag to all of the builds so that AutoQA can perform a
second round of sanity checks.

When pushed, bodhi will handle removing these tags.

Once AutoQA is good to go, then bodhi will need to ensure that no updates can
be pushed to the corresponding repos until AutoQA gives it the appropriate
karma