setup: Fix the Turbogears version requirement
When installing Bodhi, easy_install with fetch the most recent version
of the dependencies which statisfy the requirements.

The current requirement on TurboGears was:
    TurboGears >= 1.0

At the moment, the latest version of TurboGears satisfying this
requirement is 1.5, so that's what easy_install installs.

But TurboGears 1.5 requires:
    CherryPy >= 3.1.2

Whereas Bodhi requires the following:
    CherryPy<3.0

And so the installation of Bodhi failed because of this conflicts.