Only suggest candidate builds
When issueing a new update, maintainers are suggested with the latest
builds from all candidate tags.

But some of them might already have been part of another update, and
even have been pushed to stable. At this point it doesn't make sense to
still suggest them for new updates.

This commit changes that, so that Bodhi only suggest builds which are
still tagged as candidates in Koji.