models: Make the Update.critpath property a column
This makes a few things simpler.

First, we only query pkgdb at new/edit time, rather than every time we
need to know whether an update is critpath or not.

Second, it allows filtering efficiently with SQLAlchemy. (which we use
to query updates)

Now, critpath is conceptually an attribute of the packages. However,
making it an atribute of the Package model in Bodhi would not be as
interesting:

* packages have a much longer lifespan than updates in Bodhi (they
  basically stay there forever) so we'd need to have a way to update the
  critpath-yness of packages in our database.
* everything we use the critpath value for in Bodhi is related to the
  update policy, so it really relates to the Update more than the
  Package.