schemas Add the submitted_since item to ListUpdateSchema
I'm dropping the whole custom date validator in the process, which means
we lose one thing: dates in the future are not considered errors any
more.

However, I don't really see the point for our purpose: trying to list
updates submitted after a date in the future will just not return any
update, which is fine really.

Plus, the colander stuff returns a timezone-aware (UTC) datetime object,
but datetime.now() returns a timezone-naive datetime object, so the two
can't be compared.

Python < 3.2 makes it a pain to create a timezone-aware datetime object,
so I'd rather avoid it as the feature is not really needed. (for now)