Fix the \n issue in update notes
For some reason, as soon as an update is saved to the DB, its update
notes are modified: all \n characters are escaped.

It could come from a misconfiguration of SQLObject or PostgreSQL, or a
different version from what is deployed on the Fedora infrastructure,
etc...

I really can't figure out what the problem could be.

This commit fixes the symptom, not the root cause. But at least, it
allows us to use Bodhi in a reasonable way, until we can fix the actual
problem.