Update assign_alias to filter the alias in python rather than with SQL
The SQL query was relying on a group_by that we cannot get to work with
PostgreSQL as it requires all the component returned in the SELECT to be part of
the group_by. As Update.id is different for each entry in the result set
returned, having a group_by on Update.id is not uesful at all.

So what we end-up doing is getting all the alias that fits the requirements,
split them, put the year and id, as integer, in a list of tuple and we now
that the max() of this list will give us the highest year/updateid combination.