Perform a self consistency check on the mash after locking.
This is necessary to detect all sorts of race conditions.

Two things happen at the beginning of the mash:

- The update-selection process.  Updates are identified as ones that should be
  included in the mash.
- The update-locking process.  The selected updates are marked as
  'locked' so that other threads and processes (wsgi threads and
  processes) cannot modify them.

This attempts to avoid a situation where, when the mash is just getting
started, a packager will push a new update that obsoletes an update -- or
will try to unpush an update -- or will try to *delete* an update.  No
good.

The locking mechanism on updates in part prevents this, but it is
necessary to carry out this new extra sanity check *after* the updates
have been locked in order to ensure there was no interference between
the update-selection and update-locking routines.