Implement and test comment anonymization over fedmsg.
Just like #111 does for bodhi1.

A new magic map is added for all models that can be used elsewhere
called ``__anonymity_map__``: a dict that allows conditionally scrubbing
fields from JSON responses if ``anonymize`` is set to ``True``.

Currently, this is only fleshed out and tested for the ``Comment``
model, but I don't think we have a need to extend it anywhere else.