Redirect to a package's updates if it exists.
I realized when working on the bodhi1->bodhi2 redirects, that bodhi1 has
this behavior that I use all the time where you can visit
https://admin.fedoraproject.org/updates/nethack and be showns updates of
that package.  In bodhi2, this just gives you a 404, because there's no
update named just "nethack".  It's a package.

This update enhances the update_id validator to raise a 302 redirect
*if* the update_id isn't an update id but it *is* a package name.