Stop validating rpm versions.
This check is currently broken, in that it won't let you submit a build of
foo-1.2.3 for F22 if you already submitted foo-1.2.3 for F23(!)  That's crucial
functionality we need in place.

This kind of validator is something that got ported over from bodhi1 which had
years of policy baked on to it.  Now that taskotron is doing depcheck runs and
bodhi2 is using the results of those runs to gate in the mash process... we
don't even need this validator anymore, do we?