Catch other bodhi exceptions when commenting.
I think some people are hitting this when they comment
on an update that kicks it over the karma threshold, but
that threshold checking raises an exception.  It blocks the
comment from going through all together because the exception
bubbles all the way up to cornice.

This should make it so that the comment can still go through
if karma threshhold checking fails, but it will still log the
exception so we can dig more into them later.