Fix the ACLs issue for set_request and edit.
The root cause was that we have this ``validate_acls`` validator which is used
in two different ways but it was only programmed to handle one of those two.
It became critical when that first manner would fail silently if it was called
in the second manner.

I've fleshed out that validator to now handle its two different code paths and
left lots of comments about my findings.  We should get this out asap..

Longer term, we should break the validator into two different ones that handle
each role accordingly (for updates and lists of builds).