Change the mouse cursor when doing an action
Currently after pressing the button to make an update go to testing/stable
the bodhi logo (the cabbage) starts spinning.
This is really nice but easy to miss.
With this change the mouse cursor will also change, making it clearer that
something is going on.

Fixes https://github.com/fedora-infra/bodhi/issues/355