Differentiate singleton and sequence checkboxes(lists).
This fixes #397.

The bug was introduced in 2c3bba9e when we were trying to fix something else.
The bug stems from the way the list of submitted builds gets generated in
javascript before it is POSTed to the server.  It looks like this when you have
one checkbox unchecked:

```javascript
[
        "foo-1.2.3-1.fc23",
        "bar-0.1.0-1.fc23",
        false,
        "baz-900-2.fc23",
]
```

The code in this patch only pushes that ``false`` value into the list **if**
the checkbox is specially marked with a ``data-singleton`` HTML5 attribute
(which I made up).  So, now:

- lists of checkboxes should *not* get the ``false`` value
  pushed (fixing #397).
- singleton checkboxes will still get their ``false`` value
  included (keeping #348 in good shape).