masher: Don't @checkpoint the status comments or override expiration (fixes #413 & #414)
Previously we'd call Update.status_comment(), which posts the 'This update has
been pushed to testing' comment, which is unfortunately used by
Update.time_in_testing. Then something later on in the push would fail, and we
would roll back that database transaction. So now there are a bunch of updates in testing
without this status comment.