Relax checks when obsoleting older updates
Obsolete any older updates that are a subset of the newer one. There's
no need to check that the update size matches exactly; if the newer one
has a few extra builds it's still fine to obsolete the older update.

While at this, also make sure to always warn the new update submitter if
we find an update that cannot be automatically obsoleted. In this case,
warn unconditionally, even if both updates had the same submitter: they
might have forgotten about the older update.

Fixes #365