Move our RSS feeds to their own url path.
I fought 18 ways with cornice to get a simpler way to do this, but I had to end
up adding separate RSS services for each of our endpoints.

Note that, the bug here really is with firefox which sends an Accept '*/*'
header for rss feeds.  If they sent a reasonable header like
'application/rss+xml' or something like that, we could work it!  But, no.

- Fixes #493.
- Relates to #259.