markdown: Allow specifying an upstream bug tracker
Some updates will fix bug which had been reported upstream, not in
Fedora.

The update notes should still contain this valuable information, and it
should be possible for the update submitter to easily link to upstream
bug reports.

This commit lays the ground work to make this possible. It changes the
way tickets are recognized in the markdown text to:

    tracker#bugid

For example, to point to bug report number 1234 on the Fedora bugzilla,
one should now use one of the following: (they are equivalent)

    fedora#1234
    rh#1234
    rhbz#1234

And our markdown parser would automatically transform that into a link
to bug #1234 in the Red Hat bugzilla, as used for Fedora bug reports.

This also means that the current practice of only specifying the bug id
will stop working with this commit.

This is unfortunate, but it is better than having update notes
incorrectly pointing to the Red Hat bugzilla for upstream bug reports.