Restructure the way we assign aliases.
Fixes #356.

Before this change, they were ``<PREFIX>-<YEAR>-<INTEGER>``, where the last
part grows monotonically with each new update.

This commit changes that to ``<PREFIX>-<YEAR>-<HASH>``, where hash is the first
10 characters of a hash of a random uuid.

The problem with the way things are now is that:

- we have a race condition.  If you try to submit two updates at the same time
  (via a script, this is easy to do), then they both go into the code where
  their alias is assigned at the same time, they both get the same integer, and
  then when they try to commit their database transactions, one wins and the
  other fails.
- the alias-assigning code is one of the slower parts of the new update
  process.  It takes 1-2s iirc (because it has to query for the latest update
  before it and figure out its id).

The downside here is that the update aliases will no longer be meaningful like
they were.  You won't be able to look at the latest one and instantly know how
many updates there have been so far in 2015, for instance.  IMHO, this isn't
worth the trouble they're causing us.  Up for discussion!