Collapse these two emails into one.
There's no need to get two errors emails here.

Are these even actionable?  Should we further demote this to a warning?
Does it indicate a serious underlying bug that needs to be fixed?

If it's not serious, let's demote it to a warning.  Otherwise, it will
contribute to error-fatigue.  We'll stop paying attention... ;)