CLI: implent update downloading
This turned out a bit more finicky than I expected, but it does
work. Allows downloading updates by CVE, update ID, and/or build
NVR, accepting comma-separated lists for each. You can use
any combination of the three to download multiple updates at
once.

I couldn't find a better way to do 'one of these three
things must be specified' in click, please let me know if there
is one.

The old CLI added i386 and i586 --arch options for i686 systems,
but that was done back in 2009 referencing an i586 update for
F11; I'm pretty sure we don't have any i386 or i586 builds any
more.

We could add an option to set the arches to download, I guess,
but it makes the code a bit finickier and the old CLI didn't
have it, so I left it out for now.