Remove unnecessary encoding of bug titles.
Ok, so this fixes a bug that showed up first as #366, was partway fixed, but
then it showed up again as #705.

We were originally casting bug titles to a ``str`` (to ``bytes`` in python3
terminology) which caused (of course) some encoding errors.  We then addressed
this in 06476bf1 by using's kitchen's ``to_bytes`` which still forces things to
bytes, but does so in a graceful way.

The problem is that we're using mako's ``render_unicode(...)`` function at the
end of the day, which looks for things like bytes and has to force them *back*
to unicode before rendering everything out and giving it a final encoding once
done.  It's that re-decoding that caused the issue in #705.  By just keeping
everything as unicode and letting mako deal with it, this fixes #705.  (I
checked, and ``bug.title`` is indeed a ``unicode`` object).

I hacked this in place in staging to figure it out.  Compare:

- Working in stg - https://bodhi.stg.fedoraproject.org/updates/FEDORA-2015-67b99c3223
- Broken in prod - https://bodhi.fedoraproject.org/updates/FEDORA-2015-67b99c3223