Add an "agent" to all of the masher messages.
The idea here is that when a mash fails, we'll be able to easily tell *whose*
mash failed.

Then, we can change the IRC message from:

- bodhi masher failed to mash f24-updates-testing

to:

- bodhi masher failed on kevin's mash of f24-updates-testing

The benefit being that the releng team member on ``.push_duty`` will get a ping
in their IRC client when things fail so they can have a look.