Remove cornice from the Sphinx project.
The docs project currently does not build with a mysterious error
message. Removing cornice from the project seems to remove this
error message, though currently the entirety of the Sphinx project
is the cornice generated documentation. However, I am working on
using Sphinx to build a man page for the new bodhi 2 CLI, and so
the docs failing to build is a blocker for the work I need to do.

I will file an issue to re-examine cornice in the future when we
have time to determine what the issue was.