Bring the Vagrantfile into the bodhi repository.
This commit brings Ryan Lerch's Vagrantfile from
https://pagure.io/fedora-apps-vagrantfiles/ into the Bodhi main
repository. It also adjusts the Readme to be a little more thorough
and to reflect this change.

By bringing the Vagrantfile into the Bodhi repository, we will be
able to easily convert it to use Ansible for provisioning instead
of shell commands. This will help us manage the development
environment more easily in the future.