Fix the Ansible replace statement to match the sqlite default.
New Bodhi developers would not have had their development.ini
configured properly by Ansible prior to this fix. The regex failed
to match the sqlalchemy.url setting in the development.ini.example
file, and thus did not properly change that setting to connect to
postgres. This commit adjusts the regex to match the sqlite setting
and replace with postgres.