Remove locked_date_for_update and manually clear the date_locked
This function was used to clear up a mess that should never have been caused:
the case where an update is locked without a date_locked, or it is not locked
and it has a date_locked.
Instead, we now just set and unset date_locked with the locked attribute.

Since commit 354d208e3c7fe8b39903d2810befabfd0905cb7c, this code has been doing
nothing anyway, because at that point of the code, self.updates is still an empty
set: it will only be filled during the self.load_updates call later in the work()
function.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>