Rely on self.skip_mash to detect when it is ok to skip a mash.
Previously, if there had been crashes in the call to
uinfo.insert_updateinfo() the skip logic here would not run it
again which causes resumed mashes to fail. This commit removes the
logic that tries to automatically skip mashing because it assumed
that the previous mash had been successful when it had not.

The code will generally be slower now, but more reliable.
Additionally, admins will be able to explicitly skip the mash if
they know it is safe to do so.