Only show non-passing taskotron results when more than 16 builds.
Taskotron recently began to return a large number of test results
per-package. For updates with many packages, this was causing
browsers to use a whole lot of memory rendering all the results.
This commit adjusts the logic so that only non-passing results are
displayed on updates with more than 16 builds.

re #951