Use a different method to find the bodhi.server Python packages.
setuptools-28.0 changed the behavior of the exclude argument of
find_packages() such that it does not search for subpackages
inside of any excluded package. Bodhi's setup.py had been relying
on the former behavior and so it now needs to find a method that
will work with setuptools-28.0 as well as setuptools < 28.

Unfortunately, that same change also broke backwards compatibility
on the include argument to the same function, and a package's
subpackages cannot be included without including the superpackage:

https://github.com/pypa/setuptools/issues/808

To work around the above, the server packages are now searched
with
find_packages(include=['bodhi', 'bodhi.server', 'bodhi.server.*'])
and then the 'bodhi' string is removed from the resulting list.
This way we can still find bodhi.server and its subpackages
automatically in both setuptools-28.0 and setuptools < 28.

fixes #994