PEP-8 bodhi.server.models.models.
This commit also adds a new test for enforcing that the models.py
file stays compliant with PEP-8 with two exceptions: 100 character
line length and ignoring one particular flake8 error that is very
common for sqlalchemy projects. The plan is to add new modules to
this test as the code base is converted to be PEP-8 compliant over
time, until the entire code base is enforced.