approve_testing: use required testing time when update can be pushed to stable
The script currently uses the number of days that the update has been in
testing. This can lead to comments like the one seen here:

  https://bodhi.fedoraproject.org/updates/FEDORA-2016-c22fffe3da

...where the comment says the update "has reached 9 days in testing", because
the script ran 9 days after the update entered testing. For this (F24) update
it's better to say it's reached 7 days in testing, since that's the mandatory
testing time for this release.

This fixes #1017.