Allow shelldb to detect its settings from argv and document it.
This commit adjusts the small shelldb.py utility to be a little
more flexible and allow it to detect which settings files should be
used by inspecting sys.argv for any arguments that end in .ini.
Though this is a "simpleton" approach and normally I'd want
something more robust, it does work and this is just a development
tool.

Additionally, this commit documents how to use shelldb with pshell
to get an interactive Python development shell.