Send one comment when non-autokarma updates hit the threshold.
The approve_testing script did not work properly with the new
autokarma feature. For updates that had autokarma set to False, the
approve_testing script would add a bodhi comment every time it ran
to say that the update could be pushed to stable. This commit
adjusts that script, the Update.meets_testing_requirements method,
and the Update.met_testing_requirements method so that only one
comment is left.

Additionally, it was noted that there was not testing coverage on
the approve_testing script so a new test module was created to
cover the script. In order to save some testing effort, a new base
test class was created to perform some common functions.
Additionally, this commit adds several new tests on
Update.meets_testing_requirements and
Update.met_testing_requirements.

fixes #1009