Remove an unused db_factory parameter from a message consumer.
bodhi.server.consumers.updates.UpdatesHandler.__init__() accepted
a db_factory parameter that could never be used since fedmsg-hub
will not instantiate a Bodhi database for us. Most likely this was
done to aid testing or development at some point, but it would be
best to use mock for such cases going forward. This commit removes
that parameter and a resulting if statement inside the method to
simplify the initialization of the class and reduce the testing
burden.