Break the database session generator code out into a new function.
bodhi.server.consumers.updates had a copy of some code that is
repeated throughout the codebase that sets up a configuration and
a database connection. This made it trickier to add automated tests
for the surrounding code since it didn't have separated concerns.

This commit splits that code out into a new function, which will be
useful for further consolidation in future commits. That work will
be done separately so that it can be done incrementally, carefully,
and thus, slowly over time. Once this commit is in place, it will
be much easier to test bodhi.server.consumers.updates.

re #1028