mv …/functional/test_updates.py …/services/test_updates.py.
I don't personally find much use in having a distinction between
"functional" and "non-functional" tests for Bodhi. Additionally, I
find the organization structure of mapping Bodhi modules to test
modules that use the same directory heirarchy and filenames with
test_ prefixed to make it much easier to answer the question "where
are the tests for module X?".

This commit moves test_updates.py to be in
bodhi/tests/server/services/test_updates.py to follow that scheme,
since the tests in the module pertain to
bodhi.server.services.updates.