Fix a failing test.
During my last commit, I had assumed that tests didn't depend on
other tests when I shuffled them around. I discovered that one test
was failing because it was running another test to avoid copying a
line of code and a mock.

This commit adjusts the failing test to be isolated.