Drop unreachable code and add two tests to prove it's not needed.
The new_update() function (which surprisingly also handles update
edits) had some code to query for a Package and create it if it
didn't already exist. This code was unreachable, as validate_acls()
also creates the Package if it doesn't already exist, and is the
validator on new_update().

This commit drops the unreachable code and adds two tests to ensure
that new_update() correctly functions whether or not the Package is
already in the database.