Commit db transaction after creating new Builds in new_update().
This commit alters new_update() so that the database transaction
get committed after creating new Builds in new_update(). This will
solve a race condition between robosignatory signing Builds before
they are committed to the database, which should ensure that the
signed handler will find the Builds once Robosignatory emits the
message that they have been signed.

In order for the existing tests to pass, two changes were needed in
the BaseTestCase: The tests need to start and end a transaction,
and the Session needs to be removed when the test is complete.

This commit also modifies an existing test to ensure that Builds
remain after the transaction is aborted.

fixes #1111