Move fedora.client.bodhi to bodhi.client.bindings.
This commit moves the contents of fedora.client.bodhi to a new
module, bodhi.client.bindings. The code will be much more
maintainable if it is a part of Bodhi's own codebase, and
fedora.client.bodhi will be deprecated once Fedora 25 ends its
lifecycle.

This commit adds a significant number of lines of code to bodhi but
does not include tests on the code as there were no tests on
fedora.client.bodhi. This reduces the test coverage from 78% to
75%. Since the code was not tested in the source, there is no net
change in the number of lines of tested code, so this commit also
adjusts the test configuration to require a minimum 75% coverage
instead of 78%.