Raise appropriate exceptions when getting a 404 on request().
This commit adds a new UpdateNotFound Exception. This update is now
raised whenever bodhi.client.bindings.BodhiClient.request()
encounters a 404 error code on the server.

Additionally, this commit adjusts the client such that it catches
the new Exception and raises a click.BadParameter Exception so that
the user gets a nice informative error message.

fixes #1040