Delete the bodhi.spec file.
The bodhi.spec file in this repository is not in use, and is out of
date with the real spec file that is in use in Fedora's dist-git.
As there is no apparent value in keeping this spec in its current
condition, and no apparent value in maintaining it when it is not
used by any existing processes, it is best to remove it.