Only comment on non-autokarma updates meeting testing requirements.
This commit fixes an issue where non-autokarma ciritcal path
updates were getting repeat e-mails every 6 hours saying that they
could be pushed to testing. The code that does these checks is far
more complicated than it needs to be, but it was possible to fix
this issue by adding a simple and onto the end of an existing if
statement so it will do for now. I intend to refactor this area of
the code in the future.

fixes #1009