Configure the Koji client to retry
Although the Koji client has a few defaults for the options set here, I
set them all explicitly. The difference here is that ``offline_retry``
is now ``True``. It's not entirely clear from the Koji code what cases
this covers, but it should help some with the reliability problems. This
commit does not handle problems with the Kerberos server.

Fixes #1201 in part

Signed-off-by: Jeremy Cline <jeremy@jcline.org>