pip install pyramid_debugtoolbar in the Vagrant playbook.
pyramid_debugtoolbar is part of the development.ini.example. It
was recently removed from setup.py since it is not needed for
production, but this causes the Vagrant environment to fail.
Unfortunately, this package is not available in Fedora yet, so we
will pip install it for now to get the development environment
working again.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>