Fix a mocked test on PIL is which upset on EPEL 7
The test in question failed regardless of whether
b1b3bcc6f116d5f92a9735ab4fba56df1a5ab82f was applied or not and relied
on pillow not creating a new Image instance under the hood. That assertion
appears to hold true for newer versions of pillow, but not on the
version in EPEL 7.

fixes #1271

Signed-off-by: Jeremy Cline <jeremy@jcline.org>